[refactor] added sighash to transaction
