+++ bscript/interpreter/engine_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/bitcoin-sv/go-sdk/bscript"
 	"github.com/bitcoin-sv/go-sdk/bscript/interpreter/errs"
 	"github.com/bitcoin-sv/go-sdk/bscript/interpreter/scriptflag"
-	"github.com/bitcoin-sv/go-sdk/sighash"
 	"github.com/bitcoin-sv/go-sdk/transaction"
+	sighash "github.com/bitcoin-sv/go-sdk/transaction/sighash"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
