+++ transaction/signaturehash_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/bitcoin-sv/go-sdk/bscript"
-	"github.com/bitcoin-sv/go-sdk/sighash"
 	"github.com/bitcoin-sv/go-sdk/transaction"
+	sighash "github.com/bitcoin-sv/go-sdk/transaction/sighash"
 	"github.com/stretchr/testify/assert"
 )
