+++ transaction/unlocker.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/bitcoin-sv/go-sdk/bscript"
-	"github.com/bitcoin-sv/go-sdk/sighash"
+	sighash "github.com/bitcoin-sv/go-sdk/transaction/sighash"
 )
 
 // UnlockerParams params used for unlocking an input with a `bt.Unlocker`.
